requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
SQLAlchemy
python-dotenv
pydantic-settings